ocr-server/requirements.txt
Flask>=2.0.0
Flask-CORS>=3.0.0
Pillow>=8.0.0
pytesseract>=0.3.8
numpy>=1.21.0
psycopg2-binary>=2.8.6
Werkzeug>=2.0.0
